+++ src/Symfony/Component/Lock/Store/PostgreSqlStore.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\Lock\Store;
 
-use Doctrine\DBAL\Connection;
 use Symfony\Component\Lock\BlockingSharedLockStoreInterface;
 use Symfony\Component\Lock\BlockingStoreInterface;
 use Symfony\Component\Lock\Exception\InvalidArgumentException;
@@ -27,14 +26,12 @@ use Symfony\Component\Lock\PersistingStoreInterface;
  */
 class PostgreSqlStore implements BlockingSharedLockStoreInterface, BlockingStoreInterface
 {
-    private \PDO|Connection $conn;
+    private \PDO $conn;
     private string $dsn;
     private string $username = '';
     private string $password = '';
     private array $connectionOptions = [];
     private static array $storeRegistry = [];
-
-    private $dbalStore;
 
     /**
      * You can either pass an existing database connection as PDO instance or
@@ -52,23 +49,13 @@ class PostgreSqlStore implements BlockingSharedLockStoreInterface, BlockingStore
      * @throws InvalidArgumentException When PDO error mode is not PDO::ERRMODE_EXCEPTION
      * @throws InvalidArgumentException When namespace contains invalid characters
      */
-    public function __construct(\PDO|Connection|string $connOrDsn, array $options = [])
+    public function __construct(\PDO|string $connOrDsn, array $options = [])
     {
-        if ($connOrDsn instanceof Connection || (\is_string($connOrDsn) && str_contains($connOrDsn, '://'))) {
-            trigger_deprecation('symfony/lock', '5.4', 'Usage of a DBAL Connection with "%s" is deprecated and will be removed in symfony 6.0. Use "%s" instead.', __CLASS__, DoctrineDbalPostgreSqlStore::class);
-            $this->dbalStore = new DoctrineDbalPostgreSqlStore($connOrDsn);
-
-            return;
-        }
-
         if ($connOrDsn instanceof \PDO) {
             if (\PDO::ERRMODE_EXCEPTION !== $connOrDsn->getAttribute(\PDO::ATTR_ERRMODE)) {
                 throw new InvalidArgumentException(sprintf('"%s" requires PDO error mode attribute be set to throw Exceptions (i.e. $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION)).', __METHOD__));
             }
 
-            $this->conn = $connOrDsn;
-            $this->checkDriver();
-        } elseif ($connOrDsn instanceof Connection) {
             $this->conn = $connOrDsn;
             $this->checkDriver();
         } else {
@@ -82,12 +69,6 @@ class PostgreSqlStore implements BlockingSharedLockStoreInterface, BlockingStore
 
     public function save(Key $key)
     {
-        if (isset($this->dbalStore)) {
-            $this->dbalStore->save($key);
-
-            return;
-        }
-
         // prevent concurrency within the same connection
         $this->getInternalStore()->save($key);
 
@@ -110,12 +91,6 @@ class PostgreSqlStore implements BlockingSharedLockStoreInterface, BlockingStore
 
     public function saveRead(Key $key)
     {
-        if (isset($this->dbalStore)) {
-            $this->dbalStore->saveRead($key);
-
-            return;
-        }
-
         // prevent concurrency within the same connection
         $this->getInternalStore()->saveRead($key);
 
@@ -139,12 +114,6 @@ class PostgreSqlStore implements BlockingSharedLockStoreInterface, BlockingStore
 
     public function putOffExpiration(Key $key, float $ttl)
     {
-        if (isset($this->dbalStore)) {
-            $this->dbalStore->putOffExpiration($key);
-
-            return;
-        }
-
         // postgresql locks forever.
         // check if lock still exists
         if (!$this->exists($key)) {
@@ -154,12 +123,6 @@ class PostgreSqlStore implements BlockingSharedLockStoreInterface, BlockingStore
 
     public function delete(Key $key)
     {
-        if (isset($this->dbalStore)) {
-            $this->dbalStore->delete($key);
-
-            return;
-        }
-
         // Prevent deleting locks own by an other key in the same connection
         if (!$this->exists($key)) {
             return;
@@ -182,10 +145,6 @@ class PostgreSqlStore implements BlockingSharedLockStoreInterface, BlockingStore
 
     public function exists(Key $key): bool
     {
-        if (isset($this->dbalStore)) {
-            return $this->dbalStore->exists($key);
-        }
-
         $sql = "SELECT count(*) FROM pg_locks WHERE locktype='advisory' AND objid=:key AND pid=pg_backend_pid()";
         $stmt = $this->getConnection()->prepare($sql);
 
@@ -202,12 +161,6 @@ class PostgreSqlStore implements BlockingSharedLockStoreInterface, BlockingStore
 
     public function waitAndSave(Key $key)
     {
-        if (isset($this->dbalStore)) {
-            $this->dbalStore->waitAndSave($key);
-
-            return;
-        }
-
         // prevent concurrency within the same connection
         // Internal store does not allow blocking mode, because there is no way to acquire one in a single process
         $this->getInternalStore()->save($key);
@@ -224,12 +177,6 @@ class PostgreSqlStore implements BlockingSharedLockStoreInterface, BlockingStore
 
     public function waitAndSaveRead(Key $key)
     {
-        if (isset($this->dbalStore)) {
-            $this->dbalStore->waitAndSaveRead($key);
-
-            return;
-        }
-
         // prevent concurrency within the same connection
         // Internal store does not allow blocking mode, because there is no way to acquire one in a single process
         $this->getInternalStore()->saveRead($key);